+++ components/notes.tsx
@@ -1,7 +1,7 @@
 'use client';
 
 import { useEffect, useState } from 'react';
-import { supabase } from '../utils';
+import { supabase } from '../utils/supabase/client';
 
 interface Note {
   id: string;
